[TSA] fix Comparison of narrow type with wide type in loop condition
The case in Resource.cpp shouldn't ever be able to cause the infinite
loop that CodeQL is concerned about since we are removing entries from
the list (the size there is the source of the wider type)

+++ src/Resource.cpp
@@ -1011,7 +1011,7 @@ namespace D3D12TranslationLayer
     bool Resource::WaitForOutstandingResourcesIfNeeded(bool DoNotWait)
     {
         while (m_Identity &&
-               m_Identity->m_OutstandingResources.size() >= m_Identity->m_MaxOutstandingResources)
+               m_Identity->m_OutstandingResources.size() >= (size_t)m_Identity->m_MaxOutstandingResources)
         {
             auto outstandingResourceUse = m_Identity->m_OutstandingResources.front();
